feat: 共通化
