+++ modules/features/note/src/main/java/net/pantasystem/milktea/note/reaction/NoteReactionViewHelper.kt
@@ -1,16 +1,11 @@
 package net.pantasystem.milktea.note.reaction
 
 import android.content.Context
-import android.graphics.drawable.Drawable
 import android.view.View
 import android.widget.ImageView
 import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.databinding.BindingAdapter
-import com.bumptech.glide.load.DataSource
-import com.bumptech.glide.load.engine.GlideException
-import com.bumptech.glide.request.RequestListener
-import com.bumptech.glide.request.target.Target
 import dagger.hilt.android.EntryPointAccessors
 import net.pantasystem.milktea.common.glide.GlideApp
 import net.pantasystem.milktea.common_android.ui.VisibilityHelper.setMemoVisibility
@@ -56,29 +51,7 @@ object NoteReactionViewHelper {
                     }
                 }
 //                .override(min(max(reactionImageTypeView.height, 20), 120))
-                .addListener(object : RequestListener<Drawable> {
-                    override fun onLoadFailed(
-                        e: GlideException?,
-                        model: Any?,
-                        target: Target<Drawable>?,
-                        isFirstResource: Boolean,
-                    ): Boolean {
-                        return false
-                    }
-
-                    override fun onResourceReady(
-                        resource: Drawable?,
-                        model: Any?,
-                        target: Target<Drawable>?,
-                        dataSource: DataSource?,
-                        isFirstResource: Boolean,
-                    ): Boolean {
-                        resource ?: return false
-                        val imageAspectRatio: Float = resource.intrinsicWidth.toFloat() / resource.intrinsicHeight
-                        ImageAspectRatioCache.put(emoji.url ?: emoji.uri, imageAspectRatio)
-                        return false
-                    }
-                })
+                .addListener(SaveImageAspectRequestListener(emoji.url ?: emoji.uri))
                 .into(reactionImageTypeView)
         }
     }
